requirements.txt
spacy==3.7.4
transformers==4.38.1
torch==2.2.1
colorama==0.4.6
groq==0.4.2
psycopg2-binary==2.9.9
opencv-python==4.9.0.80
moviepy==1.0.3
openai-whisper==20231117
deepface==0.0.84
tf-keras==2.15.0
imageio-ffmpeg==0.4.9
streamlit==1.32.0
python-dotenv==1.0.1
setuptools==69.1.1